added user priviliges in sql

+++ _delete_device.php
@@ -13,21 +13,25 @@ if (stripos($content_type, 'application/json') === false) {
 // Include config file
 require_once "config.php";
 
-// Read the input stream
-$body = file_get_contents("php://input");
-$device_id_to_delete = json_decode($body)->device_id_to_delete;
+session_start();
+//check if they are admin to allow deleting
+$sql = "SELECT is_admin FROM users WHERE id=".$_SESSION["id"];
+$user_list = $mysqli->query($sql);
+$test = $user_list->fetch_object();
+if ($test->is_admin) {
+  // Read the input stream
+  $body = file_get_contents("php://input");
+  $device_id_to_delete = json_decode($body)->device_id_to_delete;
 
-// Display the object
-// print_r($object);
+  $sql = "DELETE FROM rtu_list WHERE rtu_id={$device_id_to_delete}";
 
-$sql = "DELETE FROM rtu_list WHERE rtu_id={$device_id_to_delete}";
-
-if ($mysqli->query($sql) === TRUE) {
-  if (mysqli_affected_rows($mysqli)) {
-    $sql = "DELETE FROM event_history WHERE rtu_id={$device_id_to_delete}";
-    if ($mysqli->query($sql) === TRUE) {
-      echo ("Successfully removed device #{$device_id_to_delete} and ".mysqli_affected_rows($mysqli)." associated data.");
-    } else echo "Error connecting: " . $mysqli->error;
-  }
-  else echo "Successfully submitted but no device was found";
-} else echo "Error connecting: " . $mysqli->error;
+  if ($mysqli->query($sql) === TRUE) {
+    if (mysqli_affected_rows($mysqli)) {
+      $sql = "DELETE FROM event_history WHERE rtu_id={$device_id_to_delete}";
+      if ($mysqli->query($sql) === TRUE) {
+        echo ("Successfully removed device #{$device_id_to_delete} and ".mysqli_affected_rows($mysqli)." associated data.");
+      } else echo "Error connecting: " . $mysqli->error;
+    }
+    else echo "Successfully submitted but no device was found";
+  } else echo "Error connecting: " . $mysqli->error;
+} else echo "Your account does not have admin privileges";
